Add contribution program

+++ anchor/programs/solstarter/src/state/contribution.rs
@@ -1,13 +1,13 @@
 use anchor_lang::prelude::*;
 
 #[account]
+#[derive(InitSpace)]
 pub struct Contribution {
-    pub amount: u32,
+    pub amount: u64,
 
     pub user_pubkey: Pubkey,
 
     pub project_pubkey: Pubkey,
 }
 
-impl Contribution {
-}
+impl Contribution {}
